Refactor ServiceState SPN Updating
Nonfunctional refactor to remove the side effects
from functions that nominally get info and send
broadcasts. This also shortens the depth of the
call chains so that it's easier to see the common
code vs the branched code. This should also make
follow-up functional changes more straightforward.

Bug: 370779140
Test: atest ServiceStateTrackerTest
Flag: EXEMPT refactor
Change-Id: I4cf99739bafa51d9a6e47d4824a1433b82e4476f