Manual network selection handling when request satellite enable
If the device is in  manual network selection mode and the user starts a satellite session then
the underlying modem will not be able to connect to the satellite network (Skylo) as Skylo is a
different network.

When a terminal attempts to use satellite communication, there is a high probability that there
will be no available networks. Therefore, the processing continued with the rest of the satellite
operations without waiting for a successful or failed response from the Automatic network
selection req.

Bug: 374301122
Flag: com.android.internal.telephony.flags.carrier_roaming_nb_iot_ntn
Fix: request from google by 'Samsung features for carrier roaming NB-IoT NTN', req 10
Test: trigger RequestSatelliteEnable with manual network selection state
Change-Id: I060a64102d9db028a7480263c503eec472a67fa3