make silence emergency nofitication broadcast protected
A user reported the Emergency Calling Wifi Notification kept
popping up and the Do Not Ask Again button was not working.

Upon inspection, the ActivityManager was blocking the broadcast
action because the broadcast was not protected.  Also, the
notification should be dismissed when clicked.

Flag: com.android.internal.telephony.flags.stop_spamming_emergency_notification
Bug: 368690328
Test: manual
Change-Id: I57c606947cac5927c16e4d4af3d18a310b51ed46