Add notification when satellite availability changes
When satellite availaibility changes, we need to notify the state to users so that they're aware of it.
Specifically, when the satellite connectivity is not available, we need to show users the reason so that
they can act accordingly. For example, if we show that satellite connectivity is not available because
location settings is disabled, users can enable location settings and check availability again.

Add getSupportedMsgApps() and change access modifier of isSatelliteServiceSupportedByCarrier() from private to public.

Test: Manual
Bug: 376694756
Flag: EXEMPT bugfix

Change-Id: Ifc70aacf8a6797e9a269715620bb7a85ed98ae20
Signed-off-by: Duke Lee <deukmo.lee@samsung.corp-partner.google.com>