Update IMS resolver/binding code to support multiuser
Instead of assuming that we will bind on the system user, improve
the ImsResolver code so that it can take into account the current
user and bind based on the active user + system user to support
HSUM. The code now also supports ImsService packages limited to
a subset of the full users on the device.

This fix also improves logging for validation purposes.

Flag: com.android.internal.telephony.flags.ims_resolver_user_aware
Bug: b/371272669
Test: atest FrameworksTelephonyTests CtsTeleponyTestCases
Test: manual testing using ImsTestApp and switching across users
Change-Id: Ica5166382e8ddc2860306b57efd2727dda133ca6