Notifying ECBM stop, if an emergency call starts during ECBM
This commit adds logic to EmergencyStateTracker to notify of ECBM stop
if an emergency call starts during ECBM.
This CL is a follow-up to the previously submitted ag/29805066.
It fixes a bug discovered during additional testing.
The bug is caused by a failure to call StopEmergencyCallbackMode when
an emergency call starts during ECBM.

Bug: 377586429
Bug: 359064059
Test: atest EmergencyStateTrackerTest
Flag: com.android.internal.telephony.flags.emergency_callback_mode_notification
Change-Id: Id18419cba24f8bb0577d41814f54d4c9bfc85aa3