Hide IDLE state from clients
Bug: 378762930
Test: SatelliteSessionControllerTest
Manual system test with Pixel Skylo
Flag: com.android.internal.telephony.flags.carrier_roaming_nb_iot_ntn

This CL also adds a new method to SatelliteSessionController that can be used to determine whether or not the IDLE state is supported. This method is used to determine whether or not the IDLE state should be hidden from apps.

Change-Id: I5318110d85b19a506783eb597d4bacdd1e07b66a