Keep ping request
Currently, if ping test fails, network validator revokes the ping
request from the target phone. However, this revoke is unnecessary if
ping is not required to pass - in which case the revoke causes
unnecessary disconnect and reconnect of internet.

Therefore, this change keeps the ping request if ping test is not
required.

Fix: 366318732
Test: manual voice call + data browsing
Flag: com.android.internal.telephony.flags.keep_ping_request

Change-Id: Ie412316ef1711978ff5d46ef76f820f7576ddb28