Update ImsNrSaModeHandler to Improve Unit Testing
Update the unit test to use an abstracted-out Phone
rather than looking at the CommandsInterface. This
abstracts the internal + stateful implementation of
managing N1 Mode from the phone so that the unit
test only tests the requests from the ImsNrSaModeHandler
rather than testing the combined behavior of multiple
callers in the Phone.

Bug: 328848947
Test: atest ImsNrSaModeHander
Flag: EXEMPT bugfix
Change-Id: I74904a93642055ae00d317fa66e42ec1753e8785