Fix crash when mcc is null in MccTable.getCountryCodeForMcc.
Previously, if SubscriptionInfo.getMccString() returns null, the method would call getCountryCodeForMcc with null, which could potentially lead to a null pointer exception. This CL modifies the method to return an empty string if the mcc is null, which is consistent with the previous implementation of countryCodeForMcc.

Note that the method has been marked as @NonNull since 2019 but the annotation is not enforced.

Bug: 394271686
Change-Id: I6426eb9d88427fc4fdd699b15518d6a5bb650f99
Test: atest
Flag: com.android.internal.telephony.flags.use_i18n_for_mcc_mapping